moved classic and whosthatpokemon components to be colocated with route
