+++ src/app/(pokemon)/classic/_components/Gamebox.tsx
