+++ src/app/(pokemon)/whosthatpokemon/_components/Gamebox.tsx
