+++ src/app/(pokemon)/whosthatpokemon/_components/StatsRange.tsx
