+++ src/components/GameWrapper.tsx
@@ -14,8 +14,8 @@ import { useAtomValue, useAtom, useSetAtom } from "jotai";
 import { useHydrateAtoms } from "jotai/utils";
 import { usePathname, useSearchParams } from "next/navigation";
 import React, { useEffect } from "react";
-import ClassicGamebox from "./classic/Gamebox";
-import WhosThatPokemonGamebox from "./whosthatpokemon/Gamebox";
+import ClassicGamebox from "@/app/(pokemon)/classic/_components/Gamebox";
+import WhosThatPokemonGamebox from "@/app/(pokemon)/whosthatpokemon/_components/Gamebox";
 import PokemonTypes from "./PokemonTypes";
 import PokemonSearch from "./PokemonSearch";
 import { getLocalTimeZone, today } from "@internationalized/date";
